Reorganize interface
Split everything into offline (kernel) and online (async) libraries.

[Api] is offline and has no Async or Core dependency.

[Api] functions return a representation of an HTTP request. This
request is then passed into a requestor function in the online
[Connection] module.

[Api] doesn't need to be functorized since requests and
response-handling are controlled externally.

Because [Api] doesn't depend on [Async], a [Lwt] implementation of the
online library would be possible.

[Retry_manager] is simplified somewhat. The interface is less general:
instead of taking in an arbitrary thunk that returns an API response, it
just takes in an [Api.t]. This is more correct insofar as it only knows
how to handle errors from the actual API.